+++ VideoApp/VideoApp/ViewControllers/Chat/ChatViewController.swift
@@ -1,0 +1,31 @@
+//
+//  Copyright (C) 2021 Twilio, Inc.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import UIKit
+
+class ChatViewController: UIViewController {
+    private let chatStore: ChatStoreWriting = ChatStore.shared
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        print("Messages: \(chatStore.messages)")
+    }
+    
+    @IBAction func doneTap(_ sender: Any) {
+        dismiss(animated: true, completion: nil)
+    }
+}
